完成简易版本客户端

+++ game/logic/changshu/majiang/card_def.go
@@ -10,7 +10,7 @@ import (
 //card_def文件写对牌的定义
 
 //四人麻将
-var fourPlayerCardDef = []uint8{
+var fourPlayerCardDef = []int32{
 	//  一万二万三万四万五万六万七八万九万
 	11, 12, 13, 14, 15, 16, 17, 18, 19,
 	11, 12, 13, 14, 15, 16, 17, 18, 19,
@@ -37,10 +37,10 @@ var fourPlayerCardDef = []uint8{
 }
 
 //二人麻将
-var threePlayerCardDef = []uint8{}
+var threePlayerCardDef = []int32{}
 
 //二人麻将
-var twoPlayerCardDef = []uint8{}
+var twoPlayerCardDef = []int32{}
 
 var (
 	log *logrus.Entry //majiang package的log
@@ -53,16 +53,16 @@ func (self *CardDef) Init(logptr *logrus.Entry) {
 	log = logptr
 }
 
-func (self *CardDef) GetBaseCard(playerCount int32) []uint8 {
-	var card []uint8
+func (self *CardDef) GetBaseCard(playerCount int32) []int32 {
+	var card []int32
 	if playerCount == 4 {
-		card = make([]uint8, len(fourPlayerCardDef))
+		card = make([]int32, len(fourPlayerCardDef))
 		copy(card, fourPlayerCardDef)
 	} else if playerCount == 3 {
-		card = make([]uint8, len(threePlayerCardDef))
+		card = make([]int32, len(threePlayerCardDef))
 		copy(card, threePlayerCardDef)
 	} else if playerCount == 2 {
-		card = make([]uint8, len(twoPlayerCardDef))
+		card = make([]int32, len(twoPlayerCardDef))
 		copy(card, twoPlayerCardDef)
 	} else {
 		log.Error("玩家人数有问题")
@@ -72,8 +72,8 @@ func (self *CardDef) GetBaseCard(playerCount int32) []uint8 {
 }
 
 //洗牌
-func (self *CardDef) RandCards(baseCard []uint8) []uint8 {
-	array := make([]uint8, len(baseCard)) //保证不会改变baseCard
+func (self *CardDef) RandCards(baseCard []int32) []int32 {
+	array := make([]int32, len(baseCard)) //保证不会改变baseCard
 	copy(array, baseCard)
 	rand.Seed(time.Now().Unix())
 	for i := len(array) - 1; i >= 0; i-- {
@@ -94,17 +94,17 @@ func (self *CardDef) randInt64(min, max int64) int64 {
 }
 
 //发牌
-func (self *CardDef) DealCard(rawcards []uint8, playercount, bankerID int32) (handCards [][]uint8, leftCards []uint8) {
-	player_cards := make([][]uint8, playercount)
+func (self *CardDef) DealCard(rawcards []int32, playercount, bankerID int32) (handCards [][]int32, leftCards []int32) {
+	player_cards := make([][]int32, playercount)
 	var leftNum = len(rawcards) //剩下的牌数量
 	for i := int32(0); i < playercount; i++ {
 		//庄家多摸一张牌
 		if i == bankerID {
-			player_cards[i] = make([]uint8, 14)
+			player_cards[i] = make([]int32, 14)
 			player_cards[i][13] = rawcards[leftNum-1]
 			leftNum--
 		} else {
-			player_cards[i] = make([]uint8, 13)
+			player_cards[i] = make([]int32, 13)
 		}
 		for index := 0; index < 13; index++ {
 			player_cards[i][index] = rawcards[leftNum-1]
@@ -117,7 +117,7 @@ func (self *CardDef) DealCard(rawcards []uint8, playercount, bankerID int32) (ha
 }
 
 //加
-func (self *CardDef) add_stack(m map[uint8]int32, card uint8) {
+func (self *CardDef) add_stack(m map[int32]int32, card int32) {
 	if _, ok := m[card]; ok {
 		m[card] = m[card] + 1
 	} else {
@@ -126,7 +126,7 @@ func (self *CardDef) add_stack(m map[uint8]int32, card uint8) {
 }
 
 //减
-func (self *CardDef) sub_stack(m map[uint8]int32, card uint8) {
+func (self *CardDef) sub_stack(m map[int32]int32, card int32) {
 	num, ok := m[card]
 	if ok == false {
 		log.Errorf("减牌%d时牌数量为0", card)
@@ -138,8 +138,8 @@ func (self *CardDef) sub_stack(m map[uint8]int32, card uint8) {
 }
 
 //统计牌数量
-func (self *CardDef) StackCards(rawcards []uint8) map[uint8]int32 {
-	var newcard = make(map[uint8]int32)
+func (self *CardDef) StackCards(rawcards []int32) map[int32]int32 {
+	var newcard = make(map[int32]int32)
 	for _, v := range rawcards {
 		self.add_stack(newcard, v)
 	}
